ClusterRules: add tests for filters and move components to Cypress library
